backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
requests
python-dotenv
pydantic
